Project layout refactoring
