+++ api/v1alpha1/envoyfleet_webhook.go
@@ -36,7 +36,7 @@ import (
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 
-	"github.com/kubeshop/kusk-gateway/cert"
+	"github.com/kubeshop/kusk-gateway/internal/cert"
 )
 
 // log is for logging in this package.
