+++ api/v1alpha1/staticroute_types.go
@@ -29,7 +29,7 @@ import (
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/kubeshop/kusk-gateway/options"
+	"github.com/kubeshop/kusk-gateway/internal/options"
 )
 
 // EDIT THIS FILE!  THIS IS SCAFFOLDING FOR YOU TO OWN!
