+++ api/v1alpha1/zz_generated.deepcopy.go
@@ -30,7 +30,7 @@ SOFTWARE.
 package v1alpha1
 
 import (
-	"github.com/kubeshop/kusk-gateway/options"
+	"github.com/kubeshop/kusk-gateway/internal/options"
 	"k8s.io/api/core/v1"
 	runtime "k8s.io/apimachinery/pkg/runtime"
 )
