+++ cmd/manager/Dockerfile-debug
@@ -1,0 +1,28 @@
+# Build the manager binary
+FROM golang:1.17 as builder
+
+WORKDIR /workspace
+# Copy the Go Modules manifests
+COPY go.mod go.mod
+COPY go.sum go.sum
+# cache deps before building and copying source so that we don't need to re-download as much
+# and so that source changes don't invalidate our downloaded layer
+RUN go mod download
+
+RUN CGO_ENABLED=0 go get github.com/go-delve/delve/cmd/dlv@v1.8.0
+
+# Copy the go source
+COPY . .
+
+# Build
+RUN --mount=type=cache,target=/root/.cache/go-build --mount=type=cache,target=/root/go/pkg CGO_ENABLED=0 GOOS=linux go build -gcflags "all=-N -l" -v -o manager cmd/manager/main.go
+
+# Use distroless as minimal base image to package the manager binary
+# Refer to https://github.com/GoogleContainerTools/distroless for more details
+FROM gcr.io/distroless/static:nonroot
+WORKDIR /
+COPY --from=builder /workspace/manager .
+COPY --from=builder /go/bin/dlv .
+USER 65532:65532
+
+ENTRYPOINT ["/dlv", "--listen=:40000", "--api-version=2", "exec", "--accept-multiclient", "--continue", "/manager"]
