+++ internal/controllers/config_manager.go
@@ -38,10 +38,10 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	gateway "github.com/kubeshop/kusk-gateway/api/v1alpha1"
-	"github.com/kubeshop/kusk-gateway/envoy/config"
-	"github.com/kubeshop/kusk-gateway/envoy/manager"
-	"github.com/kubeshop/kusk-gateway/spec"
-	"github.com/kubeshop/kusk-gateway/validation"
+	"github.com/kubeshop/kusk-gateway/internal/envoy/config"
+	"github.com/kubeshop/kusk-gateway/internal/envoy/manager"
+	"github.com/kubeshop/kusk-gateway/internal/spec"
+	"github.com/kubeshop/kusk-gateway/internal/validation"
 )
 
 const (
