+++ internal/envoy/types/cors.go
